Reorganizing repository & adding my C++ programs
All of my programs are contained in the "Vlad" directory while all of
your files are contained in the "Michael" directory. We should merge
these two and get rid of the extra level of folders at some point, but
going through each program and determining which one we'd rather keep is
too much work for me at the moment.